refactor: extract item components and history hook
