+++ client/src/App.jsx
@@ -1,24 +1,16 @@
-import { useState, useEffect, useCallback } from 'react'
-import { Line } from 'react-chartjs-2'
-import 'chart.js/auto'
-import {
-  Container,
-  Typography,
-  Box,
-  Paper,
-  Tabs,
-  Tab,
-  List,
-  ListItemButton,
-  ListItemText,
-} from '@mui/material'
+import { useState, useEffect } from 'react'
+import { Container, Typography, Box, Tabs, Tab } from '@mui/material'
+import VariationList from './components/VariationList'
+import ItemList from './components/ItemList'
+import HistoryChart from './components/HistoryChart'
+import useHistory from './hooks/useHistory'
 import './App.css'
 
 function App() {
   const [tab, setTab] = useState(0)
   const [items, setItems] = useState([])
   const [selectedItem, setSelectedItem] = useState(null)
-  const [history, setHistory] = useState([])
+  const { history } = useHistory(selectedItem)
 
   useEffect(() => {
     fetch('http://localhost:3001/api/items')
@@ -31,43 +23,9 @@ function App() {
       })
   }, [])
 
-  const fetchHistory = useCallback(async (id) => {
-    const res = await fetch(`http://localhost:3001/api/items/${id}`)
-    const json = await res.json()
-    setHistory(json)
-  }, [])
-
-  useEffect(() => {
-    if (selectedItem) {
-      fetchHistory(selectedItem)
-    }
-  }, [selectedItem, fetchHistory])
-
-  const variations = [...items]
-    .map((it) => ({
-      id: it.id,
-      variation:
-        (it.quick_status.sellPrice || 0) - (it.quick_status.buyPrice || 0),
-    }))
-    .sort((a, b) => b.variation - a.variation)
-    .slice(0, 10)
-
-  const chartData = {
-    labels: history.map((h) => new Date(h.time).toLocaleTimeString()),
-    datasets: [
-      {
-        label: 'Buy Price',
-        data: history.map((h) => h.buyPrice),
-        borderColor: 'rgb(75,192,192)',
-        fill: false,
-      },
-      {
-        label: 'Sell Price',
-        data: history.map((h) => h.sellPrice),
-        borderColor: 'rgb(192,75,75)',
-        fill: false,
-      },
-    ],
+  const handleVariationSelect = (id) => {
+    setSelectedItem(id)
+    setTab(1)
   }
 
   return (
@@ -81,53 +39,20 @@ function App() {
       </Tabs>
       {tab === 0 && (
         <Box mt={2}>
-          <List>
-            {variations.map((v) => (
-              <ListItemButton
-                key={v.id}
-                onClick={() => {
-                  setSelectedItem(v.id)
-                  setTab(1)
-                }}
-              >
-                <ListItemText
-                  primary={v.id}
-                  secondary={`Variation: ${v.variation.toFixed(2)}`}
-                />
-              </ListItemButton>
-            ))}
-          </List>
+          <VariationList items={items} onSelect={handleVariationSelect} />
         </Box>
       )}
       {tab === 1 && (
         <Box mt={2} display="flex" gap={2}>
           <Box width="30%" maxHeight={400} sx={{ overflowY: 'auto' }}>
-            <List>
-              {items.map((it) => (
-                <ListItemButton
-                  key={it.id}
-                  selected={it.id === selectedItem}
-                  onClick={() => setSelectedItem(it.id)}
-                >
-                  <ListItemText
-                    primary={it.id}
-                    secondary={`Buy: ${(it.quick_status.buyPrice || 0).toFixed(1)} Sell: ${(it.quick_status.sellPrice || 0).toFixed(1)}`}
-                  />
-                </ListItemButton>
-              ))}
-            </List>
+            <ItemList
+              items={items}
+              selectedItem={selectedItem}
+              onSelect={setSelectedItem}
+            />
           </Box>
           <Box flexGrow={1}>
-            {selectedItem && (
-              <Paper sx={{ p: 2 }}>
-                <Typography variant="h6" align="center" gutterBottom>
-                  {selectedItem}
-                </Typography>
-                <Box height={400}>
-                  <Line data={chartData} />
-                </Box>
-              </Paper>
-            )}
+            <HistoryChart history={history} selectedItem={selectedItem} />
           </Box>
         </Box>
       )}
